backend/requirements.txt
Flask
PyMySQL
PyJWT
pyotp
requests
captcha
qrcode
Flask-Cors
Flask-Limiter
Flask-SocketIO
APScheduler
Pillow